refactor: database -> storage 이름 변경
