+++ kdt-spring-voucher/src/main/java/org/prgrms/kdt/storage/VoucherStorage.java
@@ -1,10 +1,10 @@
-package org.prgrms.kdt.database;
+package org.prgrms.kdt.storage;
 
 import org.prgrms.kdt.voucher.Voucher;
 
 import java.util.List;
 
-public interface VoucherDatabase {
+public interface VoucherStorage {
     Voucher saveVoucher(Voucher newVoucher);
 
     List<Voucher> findAllVoucher();
